requirements.txt
requests
beautifulsoup4
sentence-transformers
chromadb
groq
streamlit
python-dotenv